refactor(race): judgeWinners 메서드 Winner 클래스로 이동
judgeWinners 테스트 필요해 model/Winner로 이동
